Finished restructuring conversion crate
