+++ crates/mpz-share-conversion/src/m2a/mod.rs
@@ -1,39 +1,9 @@
-use crate::ShareConversionError;
-use async_trait::async_trait;
-use mpz_common::Context;
-use mpz_fields::Field;
-use mpz_ole::{OLEeEvaluate, OLEeProvide};
-
-#[async_trait]
-impl<C: Context, F: Field, T: OLEeProvide<C, F> + Send> M2A<C, F, Provide> for T {
-    async fn convert(
-        &mut self,
-        ctx: &mut C,
-        mul_shares: Vec<F>,
-    ) -> Result<Vec<F>, ShareConversionError> {
-        let mut add_shares = self.provide(ctx, mul_shares).await?;
-        add_shares.iter_mut().for_each(|share| *share = -*share);
-
-        Ok(add_shares)
-    }
-}
-
-#[async_trait]
-impl<C: Context, F: Field, T: OLEeEvaluate<C, F> + Send> M2A<C, F, Evaluate> for T {
-    async fn convert(
-        &mut self,
-        ctx: &mut C,
-        mul_shares: Vec<F>,
-    ) -> Result<Vec<F>, ShareConversionError> {
-        self.evaluate(ctx, mul_shares)
-            .await
-            .map_err(ShareConversionError::from)
-    }
-}
+mod evaluator;
+mod provider;
 
 #[cfg(test)]
 mod tests {
-    use super::M2A;
+    use crate::M2A;
     use mpz_common::executor::test_st_executor;
     use mpz_core::{prg::Prg, Block};
     use mpz_fields::{p256::P256, UniformRand};
